requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
redis>=5.0.8
numpy>=2.0.0
faiss-cpu>=1.8.0.post1     # ← Safe default
#faiss-gpu>=1.8.0         # ← Uncomment only if you want GPU (harder to install)
pydantic>=2.8.0
httpx>=0.27.0
